Moved more shaders, cleaned up unused code and other left-over bits.
